% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSimple.R
\name{bayesConvertData}
\alias{bayesConvertData}
\title{Convert data for usage within Bayesian models.}
\usage{
bayesConvertData(df)
}
\arguments{
\item{df}{data.frame to be used for bayesian inferencing.}
}
\value{
the same data.frame with all factors converted to character.
}
\description{
Converts all columns in a data.frame that are factors to
character, except for the target column.
}
\examples{
df <- mmb::bayesConvertData(df = iris)
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
