% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSimple.R
\name{bayesInferSimple}
\alias{bayesInferSimple}
\title{This function tests for a single nominal label, using the selected features
for dependency. It retains at least one value during segmenting and then
returns the probability of the target-label (the value of the feature that
is the label). In order to determine the most likely label, this function
needs to be called once for each candidate.}
\usage{
bayesInferSimple(df, features, labelCol, selectedFeatureNames = c())
}
\arguments{
\item{df}{data.frame}

\item{features}{data.frame with bayes-features. One of the features needs to
be the label-column.}

\item{labelCol}{string with the name of the feature that represents the label.}

\item{selectedFeatureNames}{vector default \code{c()}. Vector of strings that
are the names of the features the to-predict label depends on. If an empty
vector is given, then all of the features are used (except for the label). The
order then depends on the features' order.}
}
\value{
double the probability of the target-label, using the maximum a
posteriori estimate.
}
\description{
This function tests for a single nominal label, using the selected features
for dependency. It retains at least one value during segmenting and then
returns the probability of the target-label (the value of the feature that
is the label). In order to determine the most likely label, this function
needs to be called once for each candidate.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
