% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSimple.R
\name{bayesRegressSimple}
\alias{bayesRegressSimple}
\title{This function returns the most likely numerical value for the target variable
using maximum a posteriori estimation. It retains at least two values for the
target variable during segmenting so that an empirical PDF can be estimated.}
\usage{
bayesRegressSimple(df, features, labelCol, selectedFeatureNames = c())
}
\arguments{
\item{df}{data.frame}

\item{features}{data.frame with bayes-features. One of the features needs to
be the label-column (not required or no value required).}

\item{labelCol}{string with the name of the feature that represents the label
(here the target variable for regression).}

\item{selectedFeatureNames}{vector default \code{c()}. Vector of strings that
are the names of the features the to-predict label depends on. If an empty
vector is given, then all of the features are used (except for the label). The
order then depends on the features' order.}
}
\description{
This function returns the most likely numerical value for the target variable
using maximum a posteriori estimation. It retains at least two values for the
target variable during segmenting so that an empirical PDF can be estimated.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
