% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesLatex.R
\name{bayesToLatex}
\alias{bayesToLatex}
\title{Creates a string that can be used in Latex in an e.g. equation-environment.}
\usage{
bayesToLatex(conditionalFeatures, targetFeature, includeValues = F)
}
\arguments{
\item{conditionalFeatures}{data.frame of Bayesian features, the target-
feature depends on.}

\item{targetFeature}{data.frame that holds exactly one Bayesian feature,
that is supposed to be the target-feture for Bayesian inferencing.}

\item{includeValues}{default FALSE boolean to indicate whether to include
the features' values or not, i.e. "A" vs. "A = setosa".}
}
\value{
a string that can be used in Latex documents.
}
\description{
Creates a string that can be used in Latex in an e.g. equation-environment.
}
\note{
Use \code{cat()} to print a string that can be copy-pasted.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
