% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFeature.R
\name{checkBayesFeature}
\alias{checkBayesFeature}
\title{Validate a Bayesian feature using some sanity checks.}
\usage{
checkBayesFeature(dfFeature, featName)
}
\arguments{
\item{dfFeature}{a data.frame for a single feature or variable
as constructed by @seealso \code{createFeatureForBayes}.}

\item{featName}{the name of the feature or variable of which to
obtain the value.}
}
\value{
data.frame the row corresponding to the given feature name
}
\description{
Internal function to check common arguments for function
that use samples transformed to bayes-features.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
\keyword{feature}
\keyword{internal}
