% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFeature.R
\name{createFeatureForBayes}
\alias{createFeatureForBayes}
\title{Transforms a sample's feature's value into a dataframe, that holds
its name, type and value. Currently supports numeric, factor, string
and boolean values.}
\usage{
createFeatureForBayes(name, value, isLabel = F, isDiscrete = F)
}
\arguments{
\item{name}{the name of the feature or variable.}

\item{value}{the value of the feature or variable.}

\item{isLabel}{default FALSE. Indicates whether this feature or variable
is the target variable (the label or value to predict).}

\item{isDiscrete}{default FALSE. Used to indicate whether the feature or
variable given is discrete. This will also be set to true if the value
given is a charater, factor or a logical.}
}
\value{
A data.frame with one row holding all the feature's value's
properties.
}
\description{
Transforms a sample's feature's value into a dataframe, that holds
its name, type and value. Currently supports numeric, factor, string
and boolean values.
}
\seealso{
\code{sampleToBayesFeatures} that uses this function
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
