% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRegress.R
\name{getDefaultRegressor}
\alias{getDefaultRegressor}
\title{Get the system-wide default regressor.}
\usage{
getDefaultRegressor()
}
\value{
Function the function used as the regressor. Defaults to
\code{function(data) mmb::estimatePdf(data)$argmax}.
}
\description{
Getting and setting the default regressor affects all functions
that have an overridable regressor. If this is not given, the default has
defined here will be obtained.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
