% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{getMessages}
\alias{getMessages}
\title{Get a boolean indicating whether messages are enabled system-wide.}
\usage{
getMessages()
}
\value{
Boolean to indicate whether messages are enabled or not.
}
\description{
Getter for the state of messages. Returns true if enabled.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
