% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{getWarnings}
\alias{getWarnings}
\title{Get a boolean indicating whether warnings are enabled system-wide.}
\usage{
getWarnings()
}
\value{
Boolean to indicate whether warnings are enabled or not.
}
\description{
Getter for the state of warnings. Returns true if enabled.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
