% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_make.varClosure.R
\name{make.varClosure}
\alias{make.varClosure}
\title{Creates a closure over a variable and returns its getter and setter.}
\usage{
make.varClosure(initVarVal = NULL, valValidator = NULL)
}
\arguments{
\item{initVarVal}{the initial value of the closed variable.}

\item{valValidator}{an optional function to validate each value. Function
must return boolean. If this function returns FALSE for any value, an
error is thrown.}
}
\value{
list with entries 'get' and 'set' which are getter/setter for the
variable that a closure was made over.
}
\description{
Creates a closure over a variable and returns its getter and setter.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
\keyword{internal}
