% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFeature.R
\name{sampleToBayesFeatures}
\alias{sampleToBayesFeatures}
\title{Helper function that takes one sample (e.g., a row of a dataframe
with validation data) and transforms it into a data.frame where
each row corresponds to one feature (and its value) of the sample.
This is done using @seealso \code{createFeatureForBayes}. This
operation can be thought of transposing a matrix.}
\usage{
sampleToBayesFeatures(dfRow, targetCol)
}
\arguments{
\item{dfRow}{a row of a data.frame with a value for each feature.}

\item{targetCol}{the name of the feature (column in the data.frame)
that is the target variable for classification or regression.}
}
\value{
a data.frame where the first row is the feature that
represents the label.
}
\description{
Helper function that takes one sample (e.g., a row of a dataframe
with validation data) and transforms it into a data.frame where
each row corresponds to one feature (and its value) of the sample.
This is done using @seealso \code{createFeatureForBayes}. This
operation can be thought of transposing a matrix.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
