% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRegress.R
\name{setDefaultRegressor}
\alias{setDefaultRegressor}
\title{Set a system-wide default regressor.}
\usage{
setDefaultRegressor(func)
}
\arguments{
\item{func}{a Function to use a regressor, should accept one argument,
which is a vector of numeric, and return one value, the regression.}
}
\value{
void
}
\description{
Getting and setting the default regressor affects all functions
that have an overridable regressor. If this is not given, the default has
defined here will be obtained.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
