% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{setMessages}
\alias{setMessages}
\title{Enable or disable messages system-wide.}
\usage{
setMessages(enable = TRUE)
}
\arguments{
\item{enable}{a boolean to indicate whether to enable messages or not.}
}
\value{
Boolean the state of enabled
}
\description{
Setter for enabling or disabling messages. Messages are
disabled by default. Use these to enable high verbosity.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
