% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{setWarnings}
\alias{setWarnings}
\title{Enable or disable warnings system-wide.}
\usage{
setWarnings(enable = TRUE)
}
\arguments{
\item{enable}{a boolean to indicate whether to enable warnings or not.}
}
\value{
Boolean the state of enabled
}
\description{
Setter for enabling or disabling warnings. Warnings are
enabled by default.
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
