% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSimple.R
\name{bayesProbabilitySimple}
\alias{bayesProbabilitySimple}
\title{Uses simple Bayesian inference to return the probability or relative likeli-
hood or a discrete labe or continuous value.}
\usage{
bayesProbabilitySimple(df, features, targetCol,
  selectedFeatureNames = c())
}
\arguments{
\item{df}{data.frame}

\item{features}{data.frame with bayes-features. One of the features needs to
be the label-column.}

\item{targetCol}{string with the name of the feature that represents the label.}

\item{selectedFeatureNames}{vector default \code{c()}. Vector of strings that
are the names of the features the to-predict label depends on. If an empty
vector is given, then all of the features are used (except for the label). The
order then depends on the features' order.}
}
\value{
double the probability of the target-label, using the maximum a
posteriori estimate.
}
\description{
Uses simple Bayesian inference to return the probability or relative likeli-
hood or a discrete labe or continuous value.
}
\seealso{
\code{mmb::bayesInferSimple()}
}
\author{
Sebastian Hönel \href{mailto:sebastian.honel@lnu.se}{sebastian.honel@lnu.se}
}
